refactor: Standalone Status component
